move demo image grid handling to component, away from home view
